// src/generated/models/Cr8b3_gwia_employee_exceptionsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr8b3_gwia_employee_exceptionsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr8b3_gwia_employee_exceptionsesstatecode = keyof typeof Cr8b3_gwia_employee_exceptionsesstatecode;
export const Cr8b3_gwia_employee_exceptionsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr8b3_gwia_employee_exceptionsesstatuscode = keyof typeof Cr8b3_gwia_employee_exceptionsesstatuscode;

export interface Cr8b3_gwia_employee_exceptionsesBase {
  cr8b3_gw_date?: string;
  cr8b3_gw_datetime?: string;
  "cr8b3_gw_emp_exception_parameter_id@odata.bind"?: string;
  "cr8b3_gw_emp_id@odata.bind"?: string;
  cr8b3_gw_employee_comments?: string;
  "cr8b3_gw_employee_exception_status_id@odata.bind"?: string;
  cr8b3_gw_event_date?: string;
  "cr8b3_gw_exception_parameter@odata.bind"?: string;
  cr8b3_gw_hr_comments?: string;
  cr8b3_gw_update_on?: string;
  cr8b3_gwia_employee_exceptionsid: string;
  cr8b3_name: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr8b3_gwia_employee_exceptionsesstatecode;
  statuscode?: Cr8b3_gwia_employee_exceptionsesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cr8b3_gwia_employee_exceptionses extends Cr8b3_gwia_employee_exceptionsesBase {
  cr8b3_gw_attachments?: string;
  cr8b3_gw_attachments_name?: string;
  cr8b3_gw_emp_exception_parameter_idname?: string;
  cr8b3_gw_emp_idname?: string;
  cr8b3_gw_employee_exception_status_idname?: string;
  cr8b3_gw_exception_parametername?: string;
  cr8b3_gw_month?: string;
  cr8b3_gw_year?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf46_gw_event_month?: string;
  crf46_gw_event_year?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  cr8b3_gw_emp_exception_parameter_id?: object;
  _cr8b3_gw_emp_exception_parameter_id_value?: string;
  cr8b3_gw_emp_id?: object;
  _cr8b3_gw_emp_id_value?: string;
  cr8b3_gw_employee_exception_status_id?: object;
  _cr8b3_gw_employee_exception_status_id_value?: string;
  cr8b3_gw_exception_parameter?: object;
  _cr8b3_gw_exception_parameter_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}