// src/generated/models/Cr8b3_gwia_emp_exception_parameter_mastersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr8b3_gwia_emp_exception_parameter_mastersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr8b3_gwia_emp_exception_parameter_mastersstatecode = keyof typeof Cr8b3_gwia_emp_exception_parameter_mastersstatecode;
export const Cr8b3_gwia_emp_exception_parameter_mastersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr8b3_gwia_emp_exception_parameter_mastersstatuscode = keyof typeof Cr8b3_gwia_emp_exception_parameter_mastersstatuscode;

export interface Cr8b3_gwia_emp_exception_parameter_mastersBase {
  cr8b3_gw_exception_parameter?: string;
  cr8b3_gwexceptionparaid: string;
  cr8b3_gwia_emp_exception_parameter_masterid: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr8b3_gwia_emp_exception_parameter_mastersstatecode;
  statuscode?: Cr8b3_gwia_emp_exception_parameter_mastersstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cr8b3_gwia_emp_exception_parameter_masters extends Cr8b3_gwia_emp_exception_parameter_mastersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}